+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * BREAKING CHANGE: `manufacturerDataHead` is removed from `BleDevice`
 * BREAKING CHANGE: rename `WebConfig` to `WebOptions`
 * BREAKING CHANGE: rename `ManufacturerDataFilter.data` to `ManufacturerDataFilter.payload`
+* BREAKING CHANGE: rename in-app pairing capabilities
 * Deprecation: `manufacturerData` is deprecated in BleDevice and will be removed in the future
 * Improve `scanFilter` handling
 * Use `ManufacturerData` object instead of `Uint8List` for manufacturerData