+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.14.0
+* BREAKING CHANGE: `bleDevice.name` now filters out non-printable characters
+* Add `bleDevice.rawName`
+
 ## 0.13.0
 * BREAKING CHANGE: `scanFilter` filters are now in OR relation 
 * BREAKING CHANGE: `manufacturerDataHead` is removed from `BleDevice`